Add passcode autentication for credential list
